Update the Route and now it work
