+++ Team4of5/src/App.js
@@ -8,6 +8,7 @@ import logo from './logo.svg';
 import './App.css';
 import Menu from './Team4of5_App/Menu.js';
 import Public from './Team4of5_App/Public/Public.js'
+import UserLoginSignup from './Team4of5_App/UserLoginSignup/UserLoginSignup.js'
 
 import {
   BrowserRouter as Router,
@@ -17,148 +18,146 @@ import {
   withRouter
 } from 'react-router-dom'
 
+
 const AuthExample = () => (
   <Router>
     <div>
-      <AuthButton/>
+
       <ul>
         <li><Link to="/public">Public Page</Link></li>
         <li><Link to="/login">Login Page</Link></li>
+
       </ul>
       <Route path="/public" component={Public}/>
       <Route path="/login" component={UserLoginSignup}/>
+      <Route path="/menu" component={Menu}/>
     </div>
   </Router>
 )
+//  <li><Link to="/menu">menu Page</Link></li>
 
-//Authantication Button
-const AuthButton = withRouter(({ history }) => (
-  UserLoginSignup.isAuthenticated ? (
-    <p>
-      Welcome! <button onClick={() => {
-        UserLoginSignup().signout(() => history.push('/'))
-      }}>Sign out</button>
-    </p>
-  ) : (
-    <p>You are not logged in. Please Login or Sign Up</p>
-  )
-))
+// class UserLoginSignup extends React.Component {
+//
+//   constructor(props) {
+//     super(props);
+//
+//     this.state = {
+//       email: '',
+//       password: '',
+//       formBtnTxt: 'Login',
+//       redirectToMenu: false
+//     };
+//     this.handleSubmit = this.handleSubmit.bind(this);
+//     this.switchLoginSignup = this.switchLoginSignup.bind(this);
+//   }
+//
+//
+//   handleChange(name, event) {
+//     let items = this.state;
+//     items[name] = event.target.value;
+//     this.setState(items);
+//   }
+//
+//   handleSubmit(event) {
+//     if (document.getElementById("submitBtn").value == "Login") {
+//
+//       Users.sign_in_user(this.state.email, this.state.password)
+//       .then((User) => {
+//         console.log(User);
+//         alert("Login Succeed!!");
+//           console.log('User Confirm!!');
+//           //redirect to menu if login is true
+//           this.setState({redirectToMenu: true});
+//
+//       }).catch((error) => {
+//         console.log(error);
+//         alert(error.message);
+//       });
+//     } else {
+//       Users.create_user(this.state.email, this.state.password)
+//         .then((User) => {
+//           console.log(User);
+//           alert('Sign Up Succeed!!');
+//         })
+//         .catch((error) => {
+//           console.log(error);
+//           alert(error.message);
+//         });
+//     }
+//     event.preventDefault();
+//
+//   }
+//
+//   switchLoginSignup(event) {
+//     let submitBtn = document.getElementById("submitBtn");
+//     if (submitBtn.value == "Login"){
+//       this.state.formBtnTxt = "SignUp";
+//       submitBtn.value = "SignUp";
+//     } else{
+//       submitBtn.value = "Login";
+//       this.state.formBtnTxt = "Login";
+//     }
+//   }
+//
+//   render() {
+//     const { from } = this.props.location || { from: { pathname: '/' } }
+//     const { redirectToMenu } = this.state;
+//
+//     // if (redirectToMenu) {
+//     //   console.log(from);
+//     //   return (
+//     //     <Redirect to="/Menu" push />
+//     //     //<Redirect to={from}/>
+//     //   )
+//     // }
+//   //   if (redirectToMenu) {
+//   //   return (
+//   //     <Redirect to={from}/>
+//   //   )
+//   // }
+//
+//
+//     return (
+//       <div>
+//     <Router>
+//      <div>
+//        <Route exact path="/" render={() =>(
+//          redirectToMenu ? ( <Route  component={Menu} />)
+//          : (<Route component={UserLoginSignup} />)
+//        )} />
+//
+//        </div>
+//      </Router>
+//       <form onSubmit={this.handleSubmit}>
+//         <button type="button"
+//           id="switchBtn"
+//           onClick={this.switchLoginSignup}>Login | SignUp</button>
+//
+//         <div>
+//           <label>
+//             <div>
+//               Email:
+//                     </div>
+//             <input type="text" value={this.state.email} onChange={this.handleChange.bind(this, 'email')} />
+//           </label>
+//         </div>
+//         <div>
+//           <label>
+//             <div>
+//               Password(at least 6 digits):
+//                     </div>
+//             <input type="text" value={this.state.first_name} onChange={this.handleChange.bind(this, 'password')} />
+//           </label>
+//         </div>
+//         <input type="submit"
+//           id="submitBtn"
+//           value={this.state.formBtnTxt} />
+//       </form>
+// </div>
+//     );
+//   }
+// }
 
-// PrivateRoute
-
-
-
-class UserLoginSignup extends React.Component {
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      email: '',
-      password: '',
-      formBtnTxt: 'Login',
-      redirectToMenu: false
-    };
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.switchLoginSignup = this.switchLoginSignup.bind(this);
-  }
-
-
-  handleChange(name, event) {
-    let items = this.state;
-    items[name] = event.target.value;
-    this.setState(items);
-  }
-
-  handleSubmit(event) {
-    if (document.getElementById("submitBtn").value == "Login") {
-
-      Users.sign_in_user(this.state.email, this.state.password)
-      .then((User) => {
-        console.log(User);
-        alert("Login Succeed!!");
-          console.log('User Confirm!!');
-          //redirect to menu if login is true
-          this.setState({redirectToMenu: true});
-
-      }).catch((error) => {
-        console.log(error);
-        alert(error.message);
-      });
-    } else {
-      Users.create_user(this.state.email, this.state.password)
-        .then((User) => {
-          console.log(User);
-          alert('Sign Up Succeed!!');
-        })
-        .catch((error) => {
-          console.log(error);
-          alert(error.message);
-        });
-    }
-    event.preventDefault();
-
-  }
-
-  switchLoginSignup(event) {
-    let submitBtn = document.getElementById("submitBtn");
-    if (submitBtn.value == "Login"){
-      this.state.formBtnTxt = "SignUp";
-      submitBtn.value = "SignUp";
-    } else{
-      submitBtn.value = "Login";
-      this.state.formBtnTxt = "Login";
-    }
-  }
-
-  render() {
-    const { from } = this.props.location || '/';
-    const { redirectToMenu } = this.state;
-
-    // if (redirectToMenu) {
-    //   console.log(from);
-    //   return (
-    //     <Redirect to="/Menu" push />
-    //     //<Redirect to={from}/>
-    //   )
-    // }
-    if (redirectToMenu) {
-    return (
-      <Redirect to={{pathname:'/menu', state:{from}}}/>
-    )
-  }
-
-
-    return (
-      <form onSubmit={this.handleSubmit}>
-        <button type="button"
-          id="switchBtn"
-          onClick={this.switchLoginSignup}>Login | SignUp</button>
-
-        <div>
-          <label>
-            <div>
-              Email:
-                    </div>
-            <input type="text" value={this.state.email} onChange={this.handleChange.bind(this, 'email')} />
-          </label>
-        </div>
-        <div>
-          <label>
-            <div>
-              Password(at least 6 digits):
-                    </div>
-            <input type="text" value={this.state.first_name} onChange={this.handleChange.bind(this, 'password')} />
-          </label>
-        </div>
-        <input type="submit"
-          id="submitBtn"
-          value={this.state.formBtnTxt} />
-      </form>
-
-    );
-  }
-}
 export default AuthExample
 
 // class App extends Component {
