+++ Team4of5/src/Team4of5_App/RouteMap.js
@@ -26,10 +26,10 @@ class RouteMap extends React.Component {
         <div className = 'container'>
           <Navbar/>
           <Switch>
-           <Route exact path='/setting' component={Setting} />
-           <Route exact path='/bugtrack' component={BugTracking} />
-           <Route exact path='/chat' component={Chat} />
-            <Route exact path='/menu' component={Menu} />
+          <Route exact path='/menu' component={Menu} />
+           <Route  path='/setting' component={Setting} />
+           <Route  path='/bugtrack' component={BugTracking} />
+           <Route  path='/chat' component={Chat} />
            <Route path='/projectmanage' component={ProjectManagement} />
 
           </Switch>
